requirements.txt
pyzmq
torch
rlcard